Minor Refactoring (see desc.)
- Moved SimpleHelpFormatter from utils.py to helpers.py
- Removed references to MutableField from the tests and source packages.
- Added list_field in the __all__ import.
