+++ simple_parsing/utils.py
@@ -52,59 +52,6 @@ class InconsistentArgumentError(RuntimeError):
         super().__init__(*args, **kwargs)
 
 
-class SimpleHelpFormatter(argparse.ArgumentDefaultsHelpFormatter,
-                          argparse.MetavarTypeHelpFormatter,
-                          argparse.RawDescriptionHelpFormatter):
-    """Little shorthand for using some useful HelpFormatters from argparse.
-    
-    This class inherits from argparse's `ArgumentDefaultHelpFormatter`,
-    `MetavarTypeHelpFormatter` and `RawDescriptionHelpFormatter` classes.
-
-    This produces the following resulting actions:
-    - adds a "(default: xyz)" for each argument with a default
-    - uses the name of the argument type as the metavar. For example, gives
-      "-n int" instead of "-n N" in the usage and description of the arguments.
-    - Conserves the formatting of the class and argument docstrings, if given.
-    """
-
-    def _get_default_metavar_for_optional(self, action):
-        try:
-            return super()._get_default_metavar_for_optional(action)
-        except:
-            return self._get_metavar_for_type(action.type, optional=True)
-
-
-    def _get_default_metavar_for_positional(self, action):
-        try:
-            return super()._get_default_metavar_for_positional(action)
-        except:
-            return self._get_metavar_for_type(action.type, optional=False)
-
-    def _get_metavar_for_type(self, t: Type, optional: bool=False) -> str:
-        if hasattr(t, "__name__"):
-            return t.__name__
-        elif is_union(t):
-            type_args = list(get_type_arguments(t))
-            
-            none_type = type(None)
-            while none_type in type_args:  # type: ignore
-                type_args.remove(none_type)  # type: ignore
-            
-            string = ""
-            if optional:
-                string += "["
-            middle = []
-            for t_ in type_args:
-                middle.append(self._get_metavar_for_type(t_, optional=optional))
-            string += "|".join(middle)
-            if optional:
-                string += "]"
-            return string
-        else:
-            return str(t)
-
-Formatter = SimpleHelpFormatter
-
 def camel_case(name):
     s1 = re.sub('(.)([A-Z][a-z]+)', r'\1_\2', name)
     return re.sub('([a-z0-9])([A-Z])', r'\1_\2', s1).lower()
